Conflicts:
	platform/lang-impl/src/com/intellij/psi/impl/source/codeStyle/CodeStyleManagerImpl.java

+++ platform/lang-api/src/com/intellij/util/indexing/ID.java
@@ -130,7 +130,7 @@ public class ID<K, V> {
     return myName;
   }
 
-  public long getUniqueId() {
+  public int getUniqueId() {
     return myUniqueId;
   }
